+++ R/Learner.R
@@ -24,6 +24,9 @@
 #' More (experimental) learners can be found in the GitHub repository: \url{https://github.com/mlr-org/mlr3extralearners}.
 #' A guide on how to extend \CRANpkg{mlr3} with custom learners can be found in the [mlr3book](https://mlr3book.mlr-org.com).
 #'
+#' To combine the learner with preprocessing operations like factor encoding, \CRANpkg{mlr3pipelines} is recommended.
+#' Hyperparameters stored in the `param_set` can be tuned with \CRANpkg{mlr3tuning}.
+#'
 #' @template param_id
 #' @template param_task_type
 #' @template param_param_set
@@ -55,6 +58,9 @@
 #' * `oob_error(...)`: Returns the out-of-bag error of the model as `numeric(1)`.
 #'   The learner must be tagged with property `"oob_error"`.
 #'
+#' * `loglik(...)`: Extracts the log-likelihood (c.f. [stats::logLik()]).
+#'   This can be used in measures like [mlr_measures_aic] or [mlr_measures_bic].
+#'
 #'
 #' @section Setting Hyperparameters:
 #'
@@ -77,7 +83,7 @@
 #' lrn$param_set$add(paradox::ParamFct$new("foo", levels = c("a", "b")))
 #' ```
 #'
-#' @family Learner
+#' @template seealso_learner
 #' @export
 Learner = R6Class("Learner",
   public = list(
@@ -109,7 +115,6 @@ Learner = R6Class("Learner",
     #' A complete list of candidate properties, grouped by task type, is stored in [`mlr_reflections$learner_properties`][mlr_reflections].
     properties = NULL,
 
-
     #' @field data_formats (`character()`)\cr
     #' Supported data format, e.g. `"data.table"` or `"Matrix"`.
     data_formats = NULL,
@@ -120,16 +125,27 @@ Learner = R6Class("Learner",
     #' @template field_predict_sets
     predict_sets = "test",
 
+    #' @field parallel_predict (`logical(1)`)\cr
+    #' If set to `TRUE`, use \CRANpkg{future} to calculate predictions in parallel (default: `FALSE`).
+    #' The row ids of the `task` will be split into [future::nbrOfWorkers()] chunks,
+    #' and predictions are evaluated according to the active [future::plan()].
+    #' This currently only works for methods `Learner$predict()` and `Learner$predict_newdata()`,
+    #' and has no effect during [resample()] or [benchmark()] where you have other means
+    #' to parallelize.
+    parallel_predict = FALSE,
+
     #' @field timeout (named `numeric(2)`)\cr
     #' Timeout for the learner's train and predict steps, in seconds.
     #' This works differently for different encapsulation methods, see
     #' [mlr3misc::encapsulate()].
     #' Default is `c(train = Inf, predict = Inf)`.
+    #' Also see the section on error handling the mlr3book: \url{https://mlr3book.mlr-org.com/error-handling.html}
     timeout = c(train = Inf, predict = Inf),
 
     #' @field fallback ([Learner])\cr
     #' Learner which is fitted to impute predictions in case that either the model fitting or the prediction of the top learner is not successful.
     #' Requires you to enable encapsulation, otherwise errors are not caught and the execution is terminated before the fallback learner kicks in.
+    #' Also see the section on error handling the mlr3book: \url{https://mlr3book.mlr-org.com/error-handling.html}
     fallback = NULL,
 
     #' @template field_man
@@ -139,7 +155,7 @@ Learner = R6Class("Learner",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     #'
     #' Note that this object is typically constructed via a derived classes, e.g. [LearnerClassif] or [LearnerRegr].
-    initialize = function(id, task_type, param_set = ParamSet$new(), predict_types = character(), feature_types = character(),
+    initialize = function(id, task_type, param_set = ps(), predict_types = character(), feature_types = character(),
       properties = character(), data_formats = "data.table", packages = character(), man = NA_character_) {
 
       self$id = assert_string(id, min.chars = 1L)
@@ -227,19 +243,34 @@ Learner = R6Class("Learner",
     #'
     #' @return [Prediction].
     predict = function(task, row_ids = NULL) {
+      # improve error message for the common mistake of passing a data.frame here
+      if (is.data.frame(task)) {
+        stopf("To predict on data.frames, use the method `$predict_newdata()` instead of `$predict()`")
+      }
       task = assert_task(as_task(task))
-      assert_learnable(task, self)
+      assert_predictable(task, self)
       row_ids = assert_row_ids(row_ids, null.ok = TRUE)
 
       if (is.null(self$state$model) && is.null(self$state$fallback_state$model)) {
         stopf("Cannot predict, Learner '%s' has not been trained yet", self$id)
       }
 
-      pdata = learner_predict(self, task, row_ids)
-      if (is.null(pdata))
-        return(NULL)
+      if (isTRUE(self$parallel_predict) && nbrOfWorkers() > 1L) {
+        row_ids = row_ids %??% task$row_ids
+        chunked = chunk_vector(row_ids, n_chunks = nbrOfWorkers(), shuffle = FALSE)
+        pdata = future.apply::future_lapply(chunked,
+          learner_predict, learner = self, task = task,
+          future.globals = FALSE, future.seed = TRUE)
+        pdata = do.call(c, pdata)
+      } else {
+        pdata = learner_predict(self, task, row_ids)
+      }
 
-      as_prediction(check_prediction_data(pdata))
+      if (is.null(pdata)) {
+        return(NULL)
+      } else {
+        as_prediction(check_prediction_data(pdata))
+      }
     },
 
     #' @description
@@ -261,8 +292,9 @@ Learner = R6Class("Learner",
       newdata = as.data.table(assert_data_frame(newdata, min.rows = 1L))
 
       if (is.null(task)) {
-        if (is.null(self$state$train_task))
+        if (is.null(self$state$train_task)) {
           stopf("No task stored, and no task provided")
+        }
         task = self$state$train_task$clone()
       } else {
         task = assert_task(as_task(task, clone = TRUE))
@@ -279,7 +311,7 @@ Learner = R6Class("Learner",
         # create list with correct NA types and insert it into the table newdata
         tab = task$col_info[list(impute), on = "id"]
         set(tab, j = "value", value = NA)
-        nas = set_names(pmap(tab, auto_convert), tab$id)
+        nas = set_names(pmap(tab[, !"label"], auto_convert), tab$id)
         newdata = insert_named(newdata, nas)
       }
 
@@ -300,6 +332,20 @@ Learner = R6Class("Learner",
     reset = function() {
       self$state = NULL
       invisible(self)
+    },
+
+    #' @description
+    #' Extracts the base learner from nested learner objects like
+    #' `GraphLearner` in \CRANpkg{mlr3pipelines} or `AutoTuner` in
+    #' \CRANpkg{mlr3tuning}.
+    #' Returns the [Learner] itself for regular learners.
+    #'
+    #' @param recursive (`integer(1)`)\cr
+    #'   Depth of recursion for multiple nested objects.
+    #'
+    #' @return [Learner].
+    base_learner = function(recursive = Inf) {
+      self
     }
   ),
 
@@ -349,7 +395,8 @@ Learner = R6Class("Learner",
     #' @template field_hash
     hash = function(rhs) {
       assert_ro_binding(rhs)
-      hash(class(self), self$id, self$param_set$values, private$.predict_type, self$fallback$hash)
+      calculate_hash(class(self), self$id, self$param_set$values, private$.predict_type,
+        self$fallback$hash, self$parallel_predict)
     },
 
     #' @field phash (`character(1)`)\cr
@@ -358,7 +405,7 @@ Learner = R6Class("Learner",
     #' selection (feature names).
     phash = function(rhs) {
       assert_ro_binding(rhs)
-      hash(class(self), self$id, private$.predict_type, self$fallback$hash)
+      calculate_hash(class(self), self$id, private$.predict_type, self$fallback$hash)
     },
 
     #' @field predict_type (`character(1)`)\cr
@@ -406,7 +453,10 @@ Learner = R6Class("Learner",
       switch(name,
         .param_set = value$clone(deep = TRUE),
         fallback = if (is.null(value)) NULL else value$clone(deep = TRUE),
-        state = { value$log = copy(value$log); value },
+        state = {
+          value$log = copy(value$log)
+          value
+        },
         value
       )
     }