+++ R/ResamplingHoldout.R
@@ -25,24 +25,26 @@
 #' task$filter(1:10)
 #'
 #' # Instantiate Resampling
-#' rho = rsmp("holdout", ratio = 0.5)
-#' rho$instantiate(task)
+#' holdout = rsmp("holdout", ratio = 0.5)
+#' holdout$instantiate(task)
 #'
 #' # Individual sets:
-#' rho$train_set(1)
-#' rho$test_set(1)
-#' intersect(rho$train_set(1), rho$test_set(1))
+#' holdout$train_set(1)
+#' holdout$test_set(1)
+#'
+#' # Disjunct sets:
+#' intersect(holdout$train_set(1), holdout$test_set(1))
 #'
 #' # Internal storage:
-#' rho$instance # simple list
+#' holdout$instance # simple list
 ResamplingHoldout = R6Class("ResamplingHoldout", inherit = Resampling,
   public = list(
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
-      ps = ParamSet$new(list(
-        ParamDbl$new("ratio", lower = 0, upper = 1, tags = "required")
-      ))
+      ps = ps(
+        ratio = p_dbl(0, 1, tags = "required")
+      )
       ps$values = list(ratio = 2 / 3)
 
       super$initialize(id = "holdout", param_set = ps, man = "mlr3::mlr_resamplings_holdout")
@@ -70,7 +72,8 @@ ResamplingHoldout = R6Class("ResamplingHoldout", inherit = Resampling,
 
     .combine = function(instances) {
       list(train = do.call(c, map(instances, "train")), test = do.call(c, map(instances, "test")))
-    })
+    }
+  )
 )
 
 #' @include mlr_resamplings.R