+++ R/benchmark.R
@@ -8,18 +8,9 @@
 #'   Each row defines a resampling by providing a [Task], [Learner] and an instantiated [Resampling] strategy.
 #'   The helper function [benchmark_grid()] can assist in generating an exhaustive design (see examples) and
 #'   instantiate the [Resampling]s per [Task].
-#' @param store_models (`logical(1)`)\cr
-#'   Store the fitted model in the resulting [BenchmarkResult]?
-#'   Set to `TRUE` if you want to further analyse the models or want to
-#'   extract information like variable importance.
-#' @param store_backends (`logical(1)`)\cr
-#'   Keep the [DataBackend] of the [Task] in the [BenchmarkResult]?
-#'   Set to `TRUE` if your performance measures require a [Task],
-#'   or to analyse results more conveniently.
-#'   Set to `FALSE` to reduce the file size and memory footprint
-#'   after serialization.
-#'   The current default is `TRUE`, but this eventually will be changed
-#'   in a future release.
+#' @template param_store_models
+#' @template param_store_backends
+#' @template param_encapsulate
 #'
 #' @return [BenchmarkResult].
 #'
@@ -31,6 +22,7 @@
 #' @template section_progress_bars
 #' @template section_logging
 #'
+#' @template seealso_benchmark
 #' @export
 #' @examples
 #' # benchmarking with benchmark_grid()
@@ -81,19 +73,17 @@
 #' ## Get the training set of the 2nd iteration of the featureless learner on penguins
 #' rr = bmr$aggregate()[learner_id == "classif.featureless"]$resample_result[[1]]
 #' rr$resampling$train_set(2)
-benchmark = function(design, store_models = FALSE, store_backends = TRUE) {
+benchmark = function(design, store_models = FALSE, store_backends = TRUE, encapsulate = NA_character_) {
   assert_data_frame(design, min.rows = 1L)
   assert_names(names(design), permutation.of = c("task", "learner", "resampling"))
   design$task = list(assert_tasks(as_tasks(design$task)))
   design$learner = list(assert_learners(as_learners(design$learner)))
   design$resampling = list(assert_resamplings(as_resamplings(design$resampling), instantiated = TRUE))
   assert_flag(store_models)
+  assert_flag(store_backends)
 
   # check for multiple task types
-  task_types = unique(map_chr(design$task, "task_type"))
-  if (length(task_types) > 1L) {
-    stopf("Multiple task types detected: %s", str_collapse(task_types))
-  }
+  assert_same_task_type(c(design$task, design$learner))
 
   # clone inputs
   setDT(design)
@@ -101,6 +91,9 @@ benchmark = function(design, store_models = FALSE, store_backends = TRUE) {
   design[, "task" := list(list(task[[1L]]$clone())), by = list(hashes(task))]
   design[, "learner" := list(list(learner[[1L]]$clone())), by = list(hashes(learner))]
   design[, "resampling" := list(list(resampling[[1L]]$clone())), by = list(hashes(resampling))]
+
+  # set encapsulation + fallback
+  set_encapsulation(design$learner, encapsulate)
 
   # expand the design: add rows for each resampling iteration
   grid = pmap_dtr(design, function(task, learner, resampling) {
@@ -114,7 +107,12 @@ benchmark = function(design, store_models = FALSE, store_backends = TRUE) {
   n = nrow(grid)
 
   lg$info("Running benchmark with %i resampling iterations", n)
-  pb = get_progressor(n)
+  pb = if (isNamespaceLoaded("progressr")) {
+    # NB: the progress bar needs to be created in this env
+    pb = progressr::progressor(steps = n)
+  } else {
+    NULL
+  }
 
   if (getOption("mlr3.debug", FALSE)) {
     lg$info("Running benchmark() sequentially in debug mode with %i iterations", n)